break apart DecksTable
